enhancement: rename SVG to CheckboxShape
