+++ src/lib/components/notes/CheckboxShape/CheckboxShape.style.ts
