+++ src/lib/components/notes/CheckboxShape/CheckboxShape.test.tsx
@@ -2,14 +2,14 @@ import React from 'react';
 import {cleanup, render} from '@testing-library/react';
 import {describe, expect, it, afterEach} from 'vitest';
 
-import {SVG} from './SVG';
+import {CheckboxShape} from './CheckboxShape';
 import {SVGPaths} from "@config/svg-paths";
 
 describe('<SVG />', () => {
     afterEach(cleanup);
 
     it('should render correctly with default props', () => {
-        const { container } = render(<SVG />);
+        const { container } = render(<CheckboxShape />);
         const svgElement = container.firstChild as SVGSVGElement;
 
         expect(svgElement).toBeDefined();
@@ -17,7 +17,7 @@ describe('<SVG />', () => {
     });
 
     it('should render the tick shape by default', () => {
-        const { container } = render(<SVG />);
+        const { container } = render(<CheckboxShape />);
         const polylineElement = container.querySelector('polyline');
 
         expect(polylineElement).toBeDefined();
@@ -25,7 +25,7 @@ describe('<SVG />', () => {
     });
 
     it('should render the heart shape when shape prop is "heart"', () => {
-        const { container } = render(<SVG shape="heart" />);
+        const { container } = render(<CheckboxShape shape="heart" />);
         const pathElement = container.querySelector('path');
 
         expect(pathElement).toBeDefined();
@@ -33,7 +33,7 @@ describe('<SVG />', () => {
     });
 
     it('should render the skull shape when shape prop is "skull"', () => {
-        const { container } = render(<SVG shape="skull" />);
+        const { container } = render(<CheckboxShape shape="skull" />);
         const pathElement = container.querySelector('path');
 
         expect(pathElement).toBeDefined();
